HTML version of the TSV
The words in italics are words added by the translators to help flow or
structure.  They are not malicious in design, but they wanted it clear
that those words were technically added and not part of the translation
process.  They have been wrapped in <em></em> tags, rather than square
brackets.  All words in brackets are words and passages that are in some
manuscripts in the original languages but not others.